Integrating support for querying Pull Requests
